requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.97.1
openai-agents==0.0.19
sendgrid==6.12.4
python-dotenv==1.1.1
pydantic==2.10.0
jinja2==3.1.6
python-multipart==0.0.18
Pillow==11.0.0
aiofiles==24.1.0
